Revert "Data layer refactor/cleanup (#1277)"
This reverts commit 88f767972283b48ebfe365e3bb0bca94bac6a29b.
