+++ backend/chainlit/session.py
@@ -193,9 +193,6 @@ class HTTPSession(BaseSession):
             shutil.rmtree(self.files_dir)
 
 
-ThreadQueue = Deque[tuple[Callable, object, tuple, Dict]]
-
-
 class WebsocketSession(BaseSession):
     """Internal web socket session object.
 
@@ -253,7 +250,7 @@ class WebsocketSession(BaseSession):
 
         self.restored = False
 
-        self.thread_queues: Dict[str, ThreadQueue] = {}
+        self.thread_queues = {}  # type: Dict[str, Deque[Callable]]
 
         ws_sessions_id[self.id] = self
         ws_sessions_sid[socket_id] = self
